+++ src/main/java/com/nctc2017/dao/impl/CannonDaoImpl.java
@@ -12,7 +12,6 @@ import java.util.Map.Entry;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.dao.DataAccessException;
-import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.ResultSetExtractor;
 import org.springframework.stereotype.Repository;
@@ -84,7 +83,7 @@ public class CannonDaoImpl implements CannonDao{
     }
 
 	@Override
-    public int createCannon(int cannonTemplateId) {
+    public BigDecimal createCannon(int cannonTemplateId) {
 		return executor.createCannon(cannonTemplateId);
     }
